♻️refactor : 불필요 클래스 삭제 & 디렉터리 구조 변경
