+++ src/main/java/com/waither/domain/noti/dto/request/TokenDto.java
@@ -1,4 +1,4 @@
-package com.waither.domain.noti.api.request;
+package com.waither.domain.noti.dto.request;
 
 public record TokenDto(
         String token
