+++ src/main/java/com/waither/domain/noti/repository/jpa/UserMedianRepository.java
@@ -1,6 +1,5 @@
 package com.waither.domain.noti.repository.jpa;
 
-import com.waither.domain.noti.entity.UserMedian;
 import com.waither.global.enums.Season;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
